ra8d2_sdhi_regs + ra_sdhi: switch to 32-bit registers, add full layout
Our SDHI header was modeling every register as 64-bit with an 8-byte
stride. FSP R_SDHI0_Type in R7KA8D2KF_core0.h lines 13170-13570 is
entirely 32-bit with a 4-byte stride and a different register
ordering. The old layout:

- Made SD_CMD look like a 64-bit register -- so the driver's write
of the command index went through 8-byte strobes hitting both
SD_CMD and the reserved slot at +0x04.
- Put SD_ARG at the right offset (0x08) but with double the width.
- Skipped the interleaved SD_ARG1 / SD_RSP1/3/5/7 half-word registers
entirely.
- Placed SD_INFO1 / SD_INFO2 / masks / CLK_CTRL at offsets +0x08
past their FSP positions (0x40/0x48/0x50/0x58/0x60 vs 0x38/0x3C/
0x40/0x44/0x48).
- Was missing SD_SIZE, SD_OPTION, SD_ERR_STS1/2, SD_BUF0, SDIO_MODE.

Fix:

ra8d2_sdhi_regs.h:
- All fields now uint32_t at their FSP offsets.
- Adds SD_ARG1 / SD_RSP1 / SD_RSP3 / SD_RSP5 / SD_RSP7.
- Adds SD_SIZE / SD_OPTION / SD_ERR_STS1 / SD_ERR_STS2 / SD_BUF0
/ SDIO_MODE so the full 0x6C window is documented.
- Size assertion: R_SDHI0_Type is 0x1B4 in FSP; our minimal struct
still covers 0x6C of it with correct offsets.

ra_sdhi.h / ra_sdhi.c:
- ra_sdhi_event_fn_t callback now takes uint32_t status_mask
(was uint64_t).
- ra_sdhi_get_status / ra_sdhi_clear_status take uint32_t* /
uint32_t (was uint64_t).
- send_command / set_clock read/write 32-bit values (was 64-bit
with ULL suffixes and bit-width mismatches).

test_ra_sdhi.c:
- Callback + last-mask state drop to uint32_t.
- Preset status literals drop the 0x00000000 prefix and ULL
suffix.
- stub_sdhi_cb signature now matches uint32_t.

63/63 host tests pass, cross build clean. The SDHI driver doesn't
read or write any silicon yet (still scaffold) but it'll
now talk to the right registers when the block-transfer path lands.

+++ libs/ra_hal/inc/ra_sdhi.h
@@ -39,7 +39,7 @@ extern "C" {
  * @typedef ra_sdhi_event_fn_t
  * @brief SDHI event callback.
  */
-typedef void (*ra_sdhi_event_fn_t)(void* ctx, uint8_t instance, uint64_t status_mask);
+typedef void (*ra_sdhi_event_fn_t)(void* ctx, uint8_t instance, uint32_t status_mask);
 
 /**
  * @brief Initialise an SDHI instance.
@@ -84,13 +84,13 @@ ra_sdhi_send_command(uint8_t instance, uint32_t cmd, uint32_t arg, uint32_t* out
  * @brief Read the SD_INFO1 status register.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_sdhi_get_status(uint8_t instance, uint64_t* out_mask);
+[[nodiscard]] ra_err_t ra_sdhi_get_status(uint8_t instance, uint32_t* out_mask);
 
 /**
  * @brief Clear SD_INFO1 status bits via write-0-to-clear.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_sdhi_clear_status(uint8_t instance, uint64_t mask);
+[[nodiscard]] ra_err_t ra_sdhi_clear_status(uint8_t instance, uint32_t mask);
 
 /**
  * @brief Attach an SDHI event callback (shared across instances).
